test-12/11/19
Added commands that successfully extract and run the executable for the helix client.
The parent directory extracted, has been hard coded for testing, but I will implement an os module command to bypass the need for hardcoding. Robert has provided several suggestions to look into:
 "A. Use tarfile.is_tarfile(filename) to ensure it is a valid tarfile before trying to extract (instead of checking the extension).
B. Use getmembers() on your opened tarfile and search through that list to find your directory name and/or a specific bin file you want to extract 
C. Add a path argument to your extractall() to extract to a specific place"
The purpose of this commit is so that I can begin testing the setup of environment variables for p4v and to save current progress.